electron/scripts/portrait/python/requirements.txt
transparent-background>=2.3.2
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
